add: queryやcookieで表示内容を変える
