+++ src/pages/@components/camera/camera.tsx
@@ -1,0 +1,3 @@
+export const Camera = () => {
+  return <></>;
+};
