+++ src/pages/@components/program/program.tsx
@@ -1,0 +1,10 @@
+import Link from 'next/link';
+
+export const Program = (programId: string) => {
+  return (
+    <div>
+      <h1>{programId}へようこそ</h1> {/* 企画名にすべき */}
+      <Link href="/">カメラ画面に戻る</Link>
+    </div>
+  );
+};
